+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Payments"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".list.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Payments">
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".NewCardActivity"
+            android:name=".newcard.NewCardActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Payments" />
