+++ app/src/main/java/nextstep/payments/common/component/PaymentCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.component
+package nextstep.payments.common.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -22,7 +22,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import nextstep.payments.model.Card
+import nextstep.payments.common.model.Card
 
 @Composable
 fun PaymentCard(
