+++ app/src/main/java/nextstep/payments/common/model/Card.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.model
+package nextstep.payments.common.model
 
 data class Card(
     val cardNumber: String,
