+++ app/src/main/java/nextstep/payments/list/CardListViewModel.kt
@@ -1,10 +1,11 @@
-package nextstep.payments
+package nextstep.payments.list
 
 import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import nextstep.payments.model.CardUiState
+import nextstep.payments.PaymentCardsRepository
+import nextstep.payments.list.model.CardUiState
 
 class CardListViewModel(
     private val cardsRepository: PaymentCardsRepository = PaymentCardsRepository
@@ -13,11 +14,11 @@ class CardListViewModel(
     val uiState: StateFlow<CardUiState> = _uiState.asStateFlow()
 
     fun fetchCards() {
-        val size = cardsRepository.cards.size
+        val size = PaymentCardsRepository.cards.size
         _uiState.value = when (size) {
             0 -> CardUiState.Empty
-            1 -> CardUiState.One(cardsRepository.cards.first())
-            else -> CardUiState.Many(cardsRepository.cards)
+            1 -> CardUiState.One(PaymentCardsRepository.cards.first())
+            else -> CardUiState.Many(PaymentCardsRepository.cards)
         }
     }
 }
