+++ app/src/main/java/nextstep/payments/list/MainActivity.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.list
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,7 +7,8 @@ import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.viewmodel.compose.viewModel
-import nextstep.payments.screen.CardListRoute
+import nextstep.payments.newcard.NewCardActivity
+import nextstep.payments.list.screen.CardListRoute
 import nextstep.payments.ui.theme.PaymentsTheme
 
 class MainActivity : ComponentActivity() {
